use stack instead of chainOf**** in chainOfLayerManager and LayerManager

+++ src/main/java/com/opencvtester/dataAccess/SessionFileDriver.java
@@ -8,19 +8,36 @@ import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
 
+class FileNotCreatedException extends RuntimeException{
+	private static final long serialVersionUID = 9222458802487115627L;	
+}
 public class SessionFileDriver implements SessionPersistenceDriver {
-	public void saveSession(Session session, String fileName) {
-		File file= new File(fileName);
-		if (file.exists()) {
-			file.delete();
+	public void saveSessionAs(Session session, String fileName) {
+		session.setTitle(fileName);
+		saveSession(session);
+	}
+	
+	@Override
+	public void saveSession(Session session) {
+
+		String fileName= session.getTitle();
+		if (fileName==null) {
+			FileNotCreatedException e = new FileNotCreatedException();
+			throw e;
 		}
-		try(FileOutputStream out = new FileOutputStream(fileName);ObjectOutputStream os = new ObjectOutputStream(out)) {
-			os.writeObject(session); 
-		} catch (FileNotFoundException e) {
-			e.printStackTrace();
-		} catch (IOException e) {
-			e.printStackTrace();
-		} 
+		else {
+			File file= new File(fileName);
+			if (file.exists()) {
+				file.delete();
+			}
+			try(FileOutputStream out = new FileOutputStream(fileName);ObjectOutputStream os = new ObjectOutputStream(out)) {
+				os.writeObject(session); 
+			} catch (FileNotFoundException e) {
+				e.printStackTrace();
+			} catch (IOException e) {
+				e.printStackTrace();
+			} 
+		}
 	}
 
 	@Override
@@ -38,5 +55,6 @@ public class SessionFileDriver implements SessionPersistenceDriver {
 		} 
 		return sessionTemp;	
 	}
+
 
 }
